api/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.37
asyncpg==0.30.0
psycopg[binary]==3.2.3
alembic==1.14.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
argon2-cffi==23.1.0
email-validator==2.2.0
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
aiosqlite==0.20.0
arq==0.26.3
redis==5.2.1
cryptography==44.0.0
fakeredis==2.28.1
curl-cffi==0.7.4
